+++ App.jsx
@@ -5,23 +5,15 @@ import FirstScreen from './screens/FirstScreen';
 import SecondScreen from './screens/SecondScreen';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import AppNavigator from './navigation/AppNavigator';
 
-const Stack = createNativeStackNavigator();
 
 export default function App() {
   return (
-    <NavigationContainer>
-      <Stack.Navigator>
-        <Stack.Screen 
-        name="FirstScreen" 
-        component={FirstScreen} 
-        
-        />
-      </Stack.Navigator>
-    </NavigationContainer>
+    <AppNavigator/>
   );
 }
 
 
-//基本形です
+//M_20211001
