+++ memo.txt
@@ -25,4 +25,11 @@ App.jsxで↑スタックナビゲーターを定義する
 
  ////////////////
  任意
- 
+ navigationディレクトリを作って,
+ AppNavigator.jsxを作成
+ App.jsxの中身をコピー
+ ＝＞関数で名前を変更
+
+
+ ///////////////
+ TouchableOpacity をインポートする　タッチした時の挙動をつけれる
