+++ navigation/AppNavigator.jsx
@@ -1,0 +1,26 @@
+import { StatusBar } from 'expo-status-bar';
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+import FirstScreen from '../screens/FirstScreen';
+import SecondScreen from '../screens/SecondScreen';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+
+const Stack = createNativeStackNavigator();
+
+export default AppNavigator = () => {
+    return (
+      <NavigationContainer>
+        <Stack.Navigator>
+          <Stack.Screen 
+          name="First" 
+          component={FirstScreen} 
+          />
+          <Stack.Screen 
+          name="Second" 
+          component={SecondScreen} 
+          />
+        </Stack.Navigator>
+      </NavigationContainer>
+    );
+  }
