+++ screens/FirstScreen.jsx
@@ -1,12 +1,14 @@
 import { StatusBar } from 'expo-status-bar';
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
+import { StyleSheet, Text, View,　 } from 'react-native';
 import FirstItem from '../components/FirstItem';
 
 const FirstScreen =() => {
   return (
     <View style={styles.container}>
-      <FirstItem/>
+      <FirstItem
+      onPress={() => alert('tap')}
+      />
     </View>
   );
 }
